+++ pyng/helper.py
@@ -1,0 +1,31 @@
+from pyng.space.vectors import Vector2D
+
+
+def dot_product(v1: Vector2D, v2: Vector2D):
+    return v1.x * v2.x + v1.y * v2.y
+
+
+def projection(polygon, axis):
+    """Går igenom alla vertices i ett objekt och projicerar dem på en axel, och returnerar sedan minsta och största värdena av dessa"""
+    min_proj = float("inf")
+    max_proj = float("-inf")
+    for vertex in polygon.vertices:
+        dot_product = vertex.x * axis.x + vertex.y * axis.y
+        min_proj = min(min_proj, dot_product)
+        max_proj = max(max_proj, dot_product)
+    return min_proj, max_proj
+
+
+def overlaps(projection1, projection2):
+    """Kollar om två projektioner överlappar varandra, utifrån deras minsta och största värden"""
+    return projection1[1] >= projection2[0] and projection1[0] <= projection2[1]
+
+
+def find_arithmetic_mean(vertices):
+    sum_x = 0
+    sum_y = 0
+    for v in vertices:
+        sum_x += v.x
+        sum_y += v.y
+
+    return Vector2D(sum_x / len(vertices), sum_y / len(vertices))
